feat: Install tailscale
Tailscale has to be installed here rather than in a container as it
needs to affect networking for the whole system. I had hoped to use an
apt action but it looks like that's not possible with a custom repo, so
I've instead used the code from github:RickBrewer/custom-orchid to
install it with manual shell commands.

It's interesting to me that we can use imperative commands without
putting a hash - coming from Nix this seems both extremely powerful and
to have the potential to break stuff in confusing ways in the future ...
but for installing Tailscale it's perfect as it'll always be updated