init: set image metadata
We set this up from the vanillaos template, so everything is at the "My
Custom Image" state of names, comments, etc.

Although this is currently in my personal account, once it is polished I
plan to submit it to Freshly Baked Cake. Therefore, I've set up the
metadata to have FBC emails, and a name that could fit in the FBC
convention